+++ packages/gestalt-codemods/1.35.0-1.36.0/__testfixtures__/leading-text-remove.output.js
@@ -1,0 +1,12 @@
+// @flow
+import React from 'react';
+import { Box, Text, Text as Renamed } from 'gestalt';
+
+export default function TestBox() {
+  return (
+    <Box>
+      <Renamed>Hello</Renamed>
+      <Text color='blue'>Hello</Text>
+    </Box>
+  );
+}