+++ packages/gestalt-codemods/1.35.0-1.36.0/__tests__/leading-text-remove.test.js
@@ -1,0 +1,13 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../leading-text-remove', () => {
+  return Object.assign(require.requireActual('../leading-text-remove'), {
+    parser: 'flow',
+  });
+});
+
+describe('leading-text-remove', () => {
+  ['leading-text-remove'].forEach(test => {
+    defineTest(__dirname, 'leading-text-remove', { quote: 'single' }, test);
+  });
+});